Move span exporter module
