+++ opentelemetry-sdk/src/lib.rs
@@ -16,7 +16,7 @@
 //!
 //! fn main() {
 //!     // Choose an exporter like `opentelemetry_stdout::SpanExporter`
-//!     # fn example<T: opentelemetry_sdk::export::trace::SpanExporter + 'static>(new_exporter: impl Fn() -> T) {
+//!     # fn example<T: opentelemetry_sdk::trace::SpanExporter + 'static>(new_exporter: impl Fn() -> T) {
 //!     let exporter = new_exporter();
 //!
 //!     // Create a new trace pipeline that prints to stdout
@@ -118,7 +118,6 @@
 )]
 #![cfg_attr(test, deny(warnings))]
 
-pub mod export;
 pub(crate) mod growable_array;
 
 #[cfg(feature = "logs")]
@@ -149,3 +148,4 @@ pub mod util;
 pub use resource::Resource;
 
 pub mod error;
+pub use error::ExportError;
