+++ opentelemetry-sdk/src/trace/runtime_tests.rs
@@ -2,7 +2,7 @@
 // need to run those tests one by one as the GlobalTracerProvider is a shared object between
 // threads Use cargo test -- --ignored --test-threads=1 to run those tests.
 #[cfg(any(feature = "rt-tokio", feature = "rt-tokio-current-thread"))]
-use crate::export::trace::{ExportResult, SpanExporter};
+use crate::trace::{ExportResult, SpanExporter};
 #[cfg(any(feature = "rt-tokio", feature = "rt-tokio-current-thread"))]
 use crate::runtime;
 #[cfg(any(feature = "rt-tokio", feature = "rt-tokio-current-thread"))]
@@ -30,7 +30,7 @@ struct SpanCountExporter {
 impl SpanExporter for SpanCountExporter {
     fn export(
         &mut self,
-        batch: Vec<crate::export::trace::SpanData>,
+        batch: Vec<crate::trace::SpanData>,
     ) -> BoxFuture<'static, ExportResult> {
         self.span_count.fetch_add(batch.len(), Ordering::SeqCst);
         Box::pin(async { Ok(()) })
